The grade calculations work, and some of the missing Webwork instantiation lines were filled in for both student ID formats.
The grade computation methods now work, as does the student id column management control flow for the Webwork Spreadsheet Class __init__ method.
The grade computation part is complete, and it should work for the Perm # based systems of this quarter and older quarters!

I still need to break up the computation of grades into separate methods so that curving can be done before all the categories and grades are computed. So this large, procedural method needs a refactor.

Next can come the data visualizations, including a personalized pie chart for students with a breakdown by grading category of how much of each category is filled up by their grade in that category, followed by a reorganization of the pie chart that shows how they add up to their overall grade.