Created a CurveSetter Class
This class can set curves for both individual grade items and also for category totals. For implementation, I would like to create a list of these objects, and the ones whose grade item arguments are actual grade items will curve those before the category totals are computed. Then the ones that are category totals will be computed before a total grade is computed.