Debugged and Improved Student Grade Pie Chart
The CurveSetter objects weren't actually curving when integrated into the Course's .make_master_spreadsheet() method. It wasn't making any changes at all to the grade column in question, and I finally figured out that this was because I was telling it to change an individual column (set as a variable), not that column of the overall dataframe. So I replaced the column variable with the dataframe and then [column_name] and it worked just fine.

The pie charts have lighter colors for the missing percentages in each category and no more percent labels.

There is still an bug with the master spreadsheet's max_points_hash attribute. It has over 200 possible points recorded for the homework total for one course, which is ridiculous! The max points should just be 100. I tried tracking down why this is happening but I haven't figured that out yet.